extract /jobs/remind to its own router (for now)
